# lib/api-zod/src/generated/types/createScheduleEntryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPEC-U Academic Management System API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateScheduleEntryRequest {
  teacherId: number;
  subjectId: number;
  classId: number;
  roomId: number;
  semesterId: number;
  /**
   * @minimum 1
   * @maximum 6
   */
  dayOfWeek: number;
  startTime: string;
  endTime: string;
}