# lib/api-zod/src/generated/types/scheduleEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPEC-U Academic Management System API
 * OpenAPI spec version: 1.0.0
 */

export interface ScheduleEntry {
  id: number;
  teacherId: number;
  teacherName: string;
  subjectId: number;
  subjectName: string;
  classId: number;
  className: string;
  roomId: number;
  roomName: string;
  semesterId: number;
  semesterName: string;
  /** 1=Lundi, 2=Mardi, 3=Mercredi, 4=Jeudi, 5=Vendredi, 6=Samedi */
  dayOfWeek: number;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  createdAt: Date;
}